Separate DummyLocalOscillator
