+++ tests/dummy_qrc/qm.py
@@ -1,7 +1,7 @@
 import pathlib
 
 from qibolab.channels import Channel, ChannelMap
-from qibolab.instruments.oscillator import LocalOscillator
+from qibolab.instruments.dummy import DummyLocalOscillator as LocalOscillator
 from qibolab.instruments.qm import QMSim
 from qibolab.platform import Platform
 from qibolab.serialize import (
