+++ tests/dummy_qrc/zurich.py
@@ -4,7 +4,7 @@ import pathlib
 import laboneq.simple as lo
 
 from qibolab.channels import Channel, ChannelMap
-from qibolab.instruments.oscillator import LocalOscillator
+from qibolab.instruments.dummy import DummyLocalOscillator as LocalOscillator
 from qibolab.instruments.zhinst import Zurich
 from qibolab.platform import Platform
 from qibolab.serialize import (
